+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## TBD
+
+### Enhancements
+
+* Support configuring Bugsnag through environment variables
+
 ## 1.9.0 (2021-01-05)
 
 ### Enhancements